requirements.txt
fastapi
uvicorn
requests
PyYAML
pydantic
python-multipart
